ms-rules/requirements.txt
flask
flask_restful
python-gitlab
sqlalchemy
psycopg2-binary
flask-cors